fixed issue with byte parameters not being allowed for clickable buttons, changed AR to use Messaging API in more cases

+++ AR proj/Assets/_Scripts/Prefabs.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 
 public class Prefabs : MonoBehaviour {
 
-	public enum PID : byte {
+	public enum PID : int {
 		Cube,
 		Ball,
 		Brick,
